requirements.txt
python-dotenv>=1.0
git+https://github.com/elliottech/lighter-python.git
